+++ src/main/java/com/tavisca/workshops/mgalaxy/roman/parsers/DriverClass.java
@@ -2,7 +2,6 @@ package com.tavisca.workshops.mgalaxy.roman.parsers;
 
 import com.tavisca.workshops.mgalaxy.roman.CalculateCredits;
 import com.tavisca.workshops.mgalaxy.roman.RomanNumeralCalculator;
-import com.tavisca.workshops.mgalaxy.roman.RomanToArabic;
 
 import java.util.HashMap;
 
@@ -11,50 +10,33 @@ public class DriverClass {
     public HashMap<String, String> queryToAnswer = new HashMap<>();
 
     public void processSentenceAndStoreAnswers(String query, RomanNumeralCalculator romanNumeralCalculator) {
-        CalculateCredits calculateCredits = new CalculateCredits();
-        String[] stringsOfInterest;
         Parser parser;
 
         if(query.split(" ").length == 3){
             parser = new WordsToRomanParser();
-            romanNumeralCalculator.storeItemToRomanValue(parser.parse(query));
+            parser.process(query, parser.parse(query), romanNumeralCalculator);
         }
         else if(query.startsWith("how many")){
             parser = new QuestionToWordsAndItemNameParser();
-            stringsOfInterest = parser.parse(query);
-            double creditValue = calculateCredits.calculateCredit(stringsOfInterest,romanNumeralCalculator);
-            this.queryToAnswer.put(query,stringsOfInterest[0] +  " " + stringsOfInterest[1] + " is " + getCreditValue(creditValue) + " Credits");
+            this.putQueryToAnswer(query, parser.process(query, parser.parse(query), romanNumeralCalculator));
         }
         else if(query.startsWith("how much is ")){
             parser = new QuestionToWordsParser();
-            stringsOfInterest = parser.parse(query);
-            String romanNumeral = romanNumeralCalculator.calculateRomanNumeralFromAmountInWords(stringsOfInterest[0]);
-            if(romanNumeral.length() > 0)
-                this.queryToAnswer.put(query,stringsOfInterest[0] + " is " +RomanToArabic.convert(romanNumeral));
-            else
-                this.queryToAnswer.put(query,(new UndefinedStrings()).parse(query)[0]);
+            this.putQueryToAnswer(query, parser.process(query, parser.parse(query), romanNumeralCalculator));
         }
         else if(query.endsWith("Credits")){
             parser = new WordsToCreditsParser();
-            calculateCredits = new CalculateCredits();
-            calculateCredits.calculateCreditPerItem(parser.parse(query), romanNumeralCalculator);
+            parser.process(query, parser.parse(query), romanNumeralCalculator);
         }
         else{
-            parser = new UndefinedStrings();
-            stringsOfInterest = parser.parse(query);
-            this.queryToAnswer.put(query, stringsOfInterest[0]);
+            this.putQueryToAnswer(query, "I have no idea what you are talking about");
         }
     }
 
-    public String getCreditValue(double creditValue){
-        String creditValueInString;
-        if(creditValue == (int) creditValue) creditValueInString = (int)creditValue + "";
-        else creditValueInString = creditValue + "";
 
-        return creditValueInString;
-    }
 
-   public String getQueryToAnswer(String query){
+    public String getQueryToAnswer(String query){
         return this.queryToAnswer.get(query);
     }
+    public void putQueryToAnswer(String query, String answer) { this.queryToAnswer.put(query, answer);}
 }
