+++ src/main/java/com/tavisca/workshops/mgalaxy/roman/parsers/Parser.java
@@ -3,6 +3,6 @@ package com.tavisca.workshops.mgalaxy.roman.parsers;
 import com.tavisca.workshops.mgalaxy.roman.RomanNumeralCalculator;
 
 public interface Parser {
-    public RomanNumeralCalculator romanNumeralCalculation = new RomanNumeralCalculator();
     public String[] parse(String query);
+    public String process(String query, String[] output, RomanNumeralCalculator romanNumeralCalculator);
 }
