+++ src/main/java/com/tavisca/workshops/mgalaxy/roman/parsers/WordsToCreditsParser.java
@@ -1,5 +1,8 @@
 package com.tavisca.workshops.mgalaxy.roman.parsers;
 
+import com.tavisca.workshops.mgalaxy.roman.CalculateCredits;
+import com.tavisca.workshops.mgalaxy.roman.RomanNumeralCalculator;
+
 public class WordsToCreditsParser implements Parser {
     public String[] parse(String query)  {
         String[] words = query.split(" is ");
@@ -13,5 +16,10 @@ public class WordsToCreditsParser implements Parser {
         }
         String number = words[1].split(" ")[0];
         return new String[]{amountInWords.trim(), itemName, number};
+    }
+
+    public String process(String query, String[] outputs, RomanNumeralCalculator romanNumeralCalculator){
+        (new CalculateCredits()).calculateAndStoreCreditPerItem(outputs, romanNumeralCalculator);
+        return "";
     }
 }
